[UI] Refactors clinics/projects list subtabs
Migrates clinics and projects subtabs to use separate FormListView and TaskListView widgets.
This change improves code modularity and reusability.
Removes redundant code and simplifies the structure of the subtabs.
