+++ lib/mainTabs/projects/subTabs/projects.list.sub_tab.dart
@@ -1,11 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:intl/intl.dart';
-import 'package:sgm/row_row_row_generated/tables/form.row.dart';
 import 'package:sgm/row_row_row_generated/tables/project.row.dart';
 import 'package:sgm/row_row_row_generated/tables/task.row.dart';
 import 'package:sgm/row_row_row_generated/tables/user.row.dart';
 import 'package:sgm/row_row_row_generated/tables/project_task_status.row.dart';
-import 'package:sgm/services/form.service.dart';
 import 'package:sgm/services/project.service.dart';
 import 'package:sgm/services/task.service.dart';
 import 'package:sgm/services/user.service.dart';
@@ -13,6 +11,8 @@ import 'package:sgm/services/project_task_status.service.dart';
 import 'package:sgm/widgets/paginated_data.dart';
 import 'package:sgm/widgets/task/taskview/task.view.dart';
 import 'package:sgm/widgets/task/dialog/update_task_status_dialog.dart';
+import 'package:sgm/widgets/form_list_view/form_list_view.dart';
+import 'package:sgm/widgets/task_list_view/task_list_view.dart';
 
 class ProjectsListSubTab extends StatefulWidget {
   static const String title = 'List';
@@ -30,41 +30,27 @@ abstract class ProjectsListSubTabState extends State<ProjectsListSubTab> {
 
 class _ProjectsListSubTabState extends ProjectsListSubTabState {
   ProjectRow? project;
-  final _paginatedDataKey = GlobalKey<PaginatedDataState>();
+  final _formListKey = GlobalKey<FormListViewState>();
+  final _taskListKey = GlobalKey<TaskListViewState>();
   final _userService = UserService();
   final _statusService = ProjectTaskStatusService();
-  final _formService = FormService();
-  Map<String, UserRow> _assigneeCache = {};
   Map<String, ProjectTaskStatusRow> _statusCache = {};
-  List<FormRow> _forms = [];
-  bool _isLoadingForms = true;
-
-  // Define consistent column widths as constants
-  static const double _titleWidth = 220.0;
-  static const double _statusWidth = 160.0;
-  static const double _dueDateWidth = 180.0;
-  static const double _assigneeWidth = 180.0;
-  static const double _birthdayWidth = 180.0;
-  static const double _nationalityWidth = 180.0;
-  static const double _phoneWidth = 180.0;
 
   @override
   void initState() {
     super.initState();
     _loadCaches();
-    _loadForms();
   }
 
   Future<void> _loadCaches() async {
     if (!mounted) return;
 
     try {
-      // Load all users for assignee mapping
+      // ignore: unused_local_variable
       final users = await _userService.getAllUsers(
         activated: false,
         isBanned: false,
       );
-      _assigneeCache = {for (var user in users) user.id: user};
 
       // Load all statuses for the project
       final statuses = await _statusService.getStatusByProjectID(
@@ -86,26 +72,6 @@ class _ProjectsListSubTabState extends ProjectsListSubTabState {
     }
   }
 
-  Future<void> _loadForms() async {
-    if (!mounted) return;
-    try {
-      final forms = await _formService.getFormsByProject(widget.projectId);
-      if (mounted) {
-        setState(() {
-          _forms = forms;
-          _isLoadingForms = false;
-        });
-      }
-    } catch (e) {
-      debugPrint('Error loading forms: $e');
-      if (mounted) {
-        setState(() {
-          _isLoadingForms = false;
-        });
-      }
-    }
-  }
-
   @override
   Future<void> reloadAPI() async {
     debugPrint("Reloading API");
@@ -120,35 +86,26 @@ class _ProjectsListSubTabState extends ProjectsListSubTabState {
       });
 
       // Refresh the task list
-      final paginatedState = _paginatedDataKey.currentState;
-      if (paginatedState != null) {
-        await paginatedState.refresh();
-      }
+      await reloadTasks();
     } catch (e) {
       debugPrint("Error reloading API: $e");
+    }
+  }
+
+  Future<void> reloadTasks() async {
+    final taskListState = _taskListKey.currentState;
+    if (taskListState != null) {
+      await taskListState.refresh();
     }
   }
 
   @override
   Future<void> reloadForms() async {
     if (!mounted) return;
-    setState(() => _isLoadingForms = true);
-    await _loadForms();
-  }
-
-  String _getAssigneeName(String? assigneeId) {
-    if (assigneeId == null) return '';
-    return _assigneeCache[assigneeId]?.name ?? 'Unknown';
-  }
-
-  String _getStatusName(TaskRow? row) {
-    if (row == null) return '';
-    return _statusCache[row.status]?.status ?? row.status ?? 'Unknown';
-  }
-
-  // Get status color based on status ID
-  Color _getStatusColor(String? statusId) {
-    return Colors.green;
+    final formListState = _formListKey.currentState;
+    if (formListState != null) {
+      await formListState.reload();
+    }
   }
 
   @override
@@ -169,36 +126,11 @@ class _ProjectsListSubTabState extends ProjectsListSubTabState {
               crossAxisAlignment: CrossAxisAlignment.stretch,
               children: [
                 // Forms section
-                if (_isLoadingForms)
-                  const Padding(
-                    padding: EdgeInsets.all(16.0),
-                    child: Center(child: CircularProgressIndicator()),
-                  )
-                else if (_forms.isEmpty)
-                  Padding(
-                    padding: const EdgeInsets.all(16.0),
-                    child: Text(
-                      'No forms for this project.',
-                      style: theme.textTheme.bodyMedium,
-                    ),
-                  )
-                else
-                  Column(
-                    crossAxisAlignment: CrossAxisAlignment.stretch,
-                    children: [
-                      Padding(
-                        padding: const EdgeInsets.fromLTRB(16, 16, 16, 0),
-                        child: Text(
-                          'Forms',
-                          style: theme.textTheme.titleMedium?.copyWith(
-                            fontWeight: FontWeight.bold,
-                          ),
-                        ),
-                      ),
-                      ..._forms.map((form) => _formItem(theme, form)),
-                      const Divider(height: 1),
-                    ],
-                  ),
+                FormListView(
+                  key: _formListKey,
+                  projectId: widget.projectId,
+                  onFormUpdated: reloadForms,
+                ),
 
                 // List section header
                 Container(
@@ -209,157 +141,226 @@ class _ProjectsListSubTabState extends ProjectsListSubTabState {
                 ),
                 const Divider(height: 1),
 
-                // List content
-                PaginatedData(
-                  key: _paginatedDataKey,
-                  builder: (context, data, isLoading) {
-                    // Calculate total table width from column widths
-                    final tableWidth =
-                        _titleWidth +
-                        _statusWidth +
-                        _dueDateWidth +
-                        _assigneeWidth +
-                        _birthdayWidth +
-                        _nationalityWidth +
-                        _phoneWidth;
-
-                    return SingleChildScrollView(
-                      scrollDirection: Axis.horizontal,
-                      child: SizedBox(
-                        width: tableWidth,
-                        child: Column(
-                          crossAxisAlignment: CrossAxisAlignment.start,
-                          children: [
-                            // Header row
-                            Container(
-                              decoration: BoxDecoration(
-                                color:
-                                    theme.colorScheme.surfaceContainerHighest,
-                                border: Border(
-                                  bottom: BorderSide(
-                                    color: theme.colorScheme.outlineVariant,
-                                  ),
-                                ),
-                              ),
-                              child: Row(
-                                children: [
-                                  _buildHeaderCell("Title", width: _titleWidth),
-                                  _buildHeaderCell(
-                                    "Status",
-                                    width: _statusWidth,
-                                  ),
-                                  _buildHeaderCell(
-                                    "Due Date",
-                                    width: _dueDateWidth,
-                                  ),
-                                  _buildHeaderCell(
-                                    "Assignee",
-                                    width: _assigneeWidth,
-                                  ),
-                                  _buildHeaderCell(
-                                    "Birthday",
-                                    width: _birthdayWidth,
-                                  ),
-                                  _buildHeaderCell(
-                                    "Nationality",
-                                    width: _nationalityWidth,
-                                  ),
-                                  _buildHeaderCell("Phone", width: _phoneWidth),
-                                ],
-                              ),
-                            ),
-
-                            // Data rows
-                            ...List.generate(data.length, (index) {
-                              final item = data[index] as TaskRow;
-                              return InkWell(
-                                onTap: () {
-                                  showGeneralDialog(
-                                    context: context,
-                                    pageBuilder: (context, a1, a2) {
-                                      return TaskView(task: item);
-                                    },
-                                  );
-                                },
-                                child: Container(
-                                  decoration: BoxDecoration(
-                                    border: Border(
-                                      bottom: BorderSide(
-                                        color: theme.colorScheme.outlineVariant,
-                                      ),
-                                    ),
-                                  ),
-                                  child: Row(
-                                    children: [
-                                      _buildDataCell(
-                                        item.title ?? "",
-                                        width: _titleWidth,
-                                      ),
-                                      _buildStatusCell(
-                                        item,
-                                        width: _statusWidth,
-                                      ),
-                                      _buildDataCell(
-                                        item.dateDue != null
-                                            ? _formatDateTime(item.dateDue!)
-                                            : "No Due Date",
-                                        width: _dueDateWidth,
-                                        style:
-                                            item.dateDue != null
-                                                ? null
-                                                : theme.textTheme.bodyMedium
-                                                    ?.copyWith(
-                                                      fontStyle:
-                                                          FontStyle.italic,
-                                                    ),
-                                      ),
-                                      _buildDataCell(
-                                        _getAssigneeName(item.assignee),
-                                        width: _assigneeWidth,
-                                      ),
-                                      _buildDataCell(
-                                        item.customerBirthday != null
-                                            ? _formatDateOnly(
-                                              item.customerBirthday!,
-                                            )
-                                            : "",
-                                        width: _birthdayWidth,
-                                      ),
-                                      _buildDataCell(
-                                        item.customerNationality ?? "",
-                                        width: _nationalityWidth,
-                                      ),
-                                      _buildDataCell(
-                                        item.customerPhone ?? "",
-                                        width: _phoneWidth,
-                                      ),
-                                    ],
-                                  ),
-                                ),
-                              );
-                            }),
-                          ],
-                        ),
-                      ),
-                    );
-                  },
-                  getPage: (int page, int pageSize) async {
-                    return await TaskService().getPage(
-                      widget.projectId,
-                      page,
-                      pageSize,
-                    );
-                  },
-                  getCount: () async {
-                    return await TaskService().getCount(widget.projectId);
-                  },
-                  initialPage: 1,
+                // Tasks section with standard TaskListView
+                TaskListView(
+                  key: _taskListKey,
+                  projectId: widget.projectId,
                 ),
               ],
             ),
           ),
         ),
       ],
+    );
+  }
+}
+
+class CustomTaskListView extends StatefulWidget {
+  const CustomTaskListView({
+    super.key,
+    required this.projectId,
+    required this.assigneeCache,
+    required this.statusCache,
+    required this.onStatusUpdate,
+  });
+
+  final String projectId;
+  final Map<String, UserRow> assigneeCache;
+  final Map<String, ProjectTaskStatusRow> statusCache;
+  final Future<void> Function() onStatusUpdate;
+
+  @override
+  State<CustomTaskListView> createState() => CustomTaskListViewState();
+}
+
+class CustomTaskListViewState extends State<CustomTaskListView> {
+  // Column widths
+  static const double _titleWidth = 220.0;
+  static const double _statusWidth = 160.0;
+  static const double _dueDateWidth = 180.0;
+  static const double _assigneeWidth = 180.0;
+  static const double _birthdayWidth = 180.0;
+  static const double _nationalityWidth = 180.0;
+  static const double _phoneWidth = 180.0;
+  
+  final _paginatedDataKey = GlobalKey<PaginatedDataState>();
+
+  Future<void> refresh() async {
+    final paginatedState = _paginatedDataKey.currentState;
+    if (paginatedState != null) {
+      await paginatedState.refresh();
+    }
+  }
+
+  String _getAssigneeName(String? assigneeId) {
+    if (assigneeId == null) return '';
+    return widget.assigneeCache[assigneeId]?.name ?? 'Unknown';
+  }
+
+  String _getStatusName(TaskRow? row) {
+    if (row == null) return '';
+    return widget.statusCache[row.status]?.status ?? row.status ?? 'Unknown';
+  }
+
+  // Get status color based on status ID
+  Color _getStatusColor(String? statusId) {
+    return Colors.green;
+  }
+
+  String _formatDateOnly(DateTime dateTime) {
+    // Convert UTC time to local time
+    final localDateTime = dateTime.toLocal();
+
+    // Format in "Month 12, 2020" format
+    final formattedDate = DateFormat('MMMM d, yyyy').format(localDateTime);
+
+    return formattedDate;
+  }
+
+  String _formatDateTime(DateTime dateTime) {
+    // Convert UTC time to local time
+    final localDateTime = dateTime.toLocal();
+
+    // Format in "Month 12, 2020 11:11" format with military time
+    final formattedDate = DateFormat(
+      'MMMM d, yyyy HH:mm',
+    ).format(localDateTime);
+
+    return formattedDate;
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    final theme = Theme.of(context);
+    
+    return PaginatedData(
+      key: _paginatedDataKey,
+      builder: (context, data, isLoading) {
+        // Calculate total table width from column widths
+        final tableWidth =
+            _titleWidth +
+            _statusWidth +
+            _dueDateWidth +
+            _assigneeWidth +
+            _birthdayWidth +
+            _nationalityWidth +
+            _phoneWidth;
+
+        return SingleChildScrollView(
+          scrollDirection: Axis.horizontal,
+          child: SizedBox(
+            width: tableWidth,
+            child: Column(
+              crossAxisAlignment: CrossAxisAlignment.start,
+              children: [
+                // Header row
+                Container(
+                  decoration: BoxDecoration(
+                    color: theme.colorScheme.surfaceContainerHighest,
+                    border: Border(
+                      bottom: BorderSide(
+                        color: theme.colorScheme.outlineVariant,
+                      ),
+                    ),
+                  ),
+                  child: Row(
+                    children: [
+                      _buildHeaderCell("Title", width: _titleWidth),
+                      _buildHeaderCell("Status", width: _statusWidth),
+                      _buildHeaderCell("Due Date", width: _dueDateWidth),
+                      _buildHeaderCell("Assignee", width: _assigneeWidth),
+                      _buildHeaderCell("Birthday", width: _birthdayWidth),
+                      _buildHeaderCell("Nationality", width: _nationalityWidth),
+                      _buildHeaderCell("Phone", width: _phoneWidth),
+                    ],
+                  ),
+                ),
+
+                // Data rows
+                ...List.generate(data.length, (index) {
+                  final item = data[index] as TaskRow;
+                  return InkWell(
+                    onTap: () {
+                      showGeneralDialog(
+                        context: context,
+                        pageBuilder: (context, a1, a2) {
+                          return TaskView(task: item);
+                        },
+                      );
+                    },
+                    child: Container(
+                      decoration: BoxDecoration(
+                        border: Border(
+                          bottom: BorderSide(
+                            color: theme.colorScheme.outlineVariant,
+                          ),
+                        ),
+                      ),
+                      child: Row(
+                        children: [
+                          _buildDataCell(
+                            item.title ?? "",
+                            width: _titleWidth,
+                          ),
+                          _buildStatusCell(
+                            item,
+                            width: _statusWidth,
+                          ),
+                          _buildDataCell(
+                            item.dateDue != null
+                                ? _formatDateTime(item.dateDue!)
+                                : "No Due Date",
+                            width: _dueDateWidth,
+                            style:
+                                item.dateDue != null
+                                    ? null
+                                    : theme.textTheme.bodyMedium
+                                        ?.copyWith(
+                                          fontStyle:
+                                              FontStyle.italic,
+                                        ),
+                          ),
+                          _buildDataCell(
+                            _getAssigneeName(item.assignee),
+                            width: _assigneeWidth,
+                          ),
+                          _buildDataCell(
+                            item.customerBirthday != null
+                                ? _formatDateOnly(
+                                  item.customerBirthday!,
+                                )
+                                : "",
+                            width: _birthdayWidth,
+                          ),
+                          _buildDataCell(
+                            item.customerNationality ?? "",
+                            width: _nationalityWidth,
+                          ),
+                          _buildDataCell(
+                            item.customerPhone ?? "",
+                            width: _phoneWidth,
+                          ),
+                        ],
+                      ),
+                    ),
+                  );
+                }),
+              ],
+            ),
+          ),
+        );
+      },
+      getPage: (int page, int pageSize) async {
+        return await TaskService().getPage(
+          widget.projectId,
+          page,
+          pageSize,
+        );
+      },
+      getCount: () async {
+        return await TaskService().getCount(widget.projectId);
+      },
+      initialPage: 1,
     );
   }
 
@@ -396,13 +397,12 @@ class _ProjectsListSubTabState extends ProjectsListSubTabState {
         onTap: () {
           showDialog(
             context: context,
-            builder:
-                (context) => UpdateTaskStatusDialog(
-                  projectId: widget.projectId,
-                  taskId: row.id,
-                  currentStatus: row.status ?? '',
-                  onStatusUpdated: reloadAPI,
-                ),
+            builder: (context) => UpdateTaskStatusDialog(
+              projectId: widget.projectId,
+              taskId: row.id,
+              currentStatus: row.status ?? '',
+              onStatusUpdated: widget.onStatusUpdate,
+            ),
           );
         },
         child: Align(
@@ -427,56 +427,6 @@ class _ProjectsListSubTabState extends ProjectsListSubTabState {
             ),
           ),
         ),
-      ),
-    );
-  }
-
-  String _formatDateOnly(DateTime dateTime) {
-    // Convert UTC time to local time
-    final localDateTime = dateTime.toLocal();
-
-    // Format in "Month 12, 2020" format
-    final formattedDate = DateFormat('MMMM d, yyyy').format(localDateTime);
-
-    return formattedDate;
-  }
-
-  String _formatDateTime(DateTime dateTime) {
-    // Convert UTC time to local time
-    final localDateTime = dateTime.toLocal();
-
-    // Format in "Month 12, 2020 11:11" format with military time
-    final formattedDate = DateFormat(
-      'MMMM d, yyyy HH:mm',
-    ).format(localDateTime);
-
-    return formattedDate;
-  }
-
-  Widget _formItem(ThemeData theme, FormRow form) {
-    return Container(
-      padding: const EdgeInsets.fromLTRB(16, 8, 16, 8),
-      decoration: BoxDecoration(
-        border: Border(
-          bottom: BorderSide(color: theme.colorScheme.outlineVariant),
-        ),
-      ),
-      child: Row(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          Icon(Icons.description, color: theme.colorScheme.primary),
-          const SizedBox(width: 8),
-          Column(
-            crossAxisAlignment: CrossAxisAlignment.start,
-            children: [
-              Text(
-                form.name ?? 'Untitled Form',
-                style: theme.textTheme.titleMedium,
-              ),
-              Text('${form.description}'),
-            ],
-          ),
-        ],
       ),
     );
   }
