+++ lib/widgets/calendar_view/calendar_view.dart
@@ -1,0 +1,10 @@
+import 'package:flutter/material.dart';
+
+class CalendarView extends StatelessWidget {
+  const CalendarView({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return const Placeholder();
+  }
+}
